chore: add Sentry to the project
This commit adds Sentry to the project. It includes changes to the
.gitignore file to ignore the .sentryclirc file, a new sentry client
configuration file, a new sentry server configuration file, a new
sentry properties file, and changes to the next.config.mjs file to
configure Sentry. It also includes changes to the layout.tsx file to
remove the Highlight component and add the SentryIdentifyUser
component, a new sentry.tsx file to identify the user in Sentry, and
a new _error.js file to handle errors.